Move custom cop into lib
