+++ .rubocop.yml
@@ -1,5 +1,5 @@
 inherit_from: .rubocop_todo.yml
-require: ./haiku_comments.rb
+require: ./lib/rubocop/cop/lint/haiku_comments.rb
 
 AllCops:
   NewCops: enable
